プレイヤーの基本操作と道具操作処理のスクリプト分ける
