+++ BooFarmPro/ProjectSettings/EditorBuildSettings.asset
@@ -8,4 +8,7 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/Main.unity
     guid: 2cda990e2423bbf4892e6590ba056729
+  - enabled: 1
+    path: Assets/Scenes/Home.unity
+    guid: 8794f09979278cf488e6f4f44531db9c
   m_configObjects: {}
